5th task:errors fixed

+++ recursion/5-sqrt_recursion.c
@@ -1,18 +1,17 @@
 #include "main.h"
 /**
- * _sqrt_recursion - returns the natural square root of a number
  * _sqrt_helper - finds square root
+ * @low: lower bound of the search range
+ * @high: upper bound of the search range
  * @n: number
- * @low: equals 1
- * @high: equals n
  *
  * Return: int
  */
 
 int _sqrt_helper(int n, int low, int high)
 {
-	int mid;
-	int square;
+	double mid;
+	double square;
 
 	if (low > high)
 		return (-1);
@@ -27,6 +26,13 @@ int _sqrt_helper(int n, int low, int high)
 	else
 		return (_sqrt_helper(n, low, mid - 1));
 }
+
+/**
+ * _sqrt_recursion - returns the natural square root of a number
+ * @n: number
+ *
+ * Return: int
+ */
 
 int _sqrt_recursion(int n)
 {
